fixed the tests and refactored
