+++ src/helper/streaming/stream.test.ts
@@ -1,4 +1,5 @@
 import { Context } from '../../context'
+import { isOldBunVersion } from './utils'
 import { stream } from '.'
 
 describe('Basic Streaming Helper', () => {
@@ -47,6 +48,11 @@ describe('Basic Streaming Helper', () => {
   })
 
   it('Check stream Response if aborted by abort signal', async () => {
+    // Emulate an old version of Bun (version 1.1.0) for this specific test case
+    // @ts-expect-error Bun is not typed
+    global.Bun = {
+      version: '1.1.0',
+    }
     const ac = new AbortController()
     const req = new Request('http://localhost/', { signal: ac.signal })
     const c = new Context(req)
@@ -69,9 +75,16 @@ describe('Basic Streaming Helper', () => {
     expect(value).toEqual(new Uint8Array([0]))
     ac.abort()
     expect(aborted).toBeTruthy()
+    // @ts-expect-error Bun is not typed
+    delete global.Bun
   })
 
   it('Check stream Response if pipe is aborted by abort signal', async () => {
+    // Emulate an old version of Bun (version 1.1.0) for this specific test case
+    // @ts-expect-error Bun is not typed
+    global.Bun = {
+      version: '1.1.0',
+    }
     const ac = new AbortController()
     const req = new Request('http://localhost/', { signal: ac.signal })
     const c = new Context(req)
@@ -91,6 +104,8 @@ describe('Basic Streaming Helper', () => {
     ac.abort()
     await pReading
     expect(aborted).toBeTruthy()
+    // @ts-expect-error Bun is not typed
+    delete global.Bun
   })
 
   it('Check stream Response if error occurred', async () => {
